Improved cloudflare secret handling script

+++ ops/cloudflare-secret.sh
@@ -1,14 +1,28 @@
 #!/usr/bin
-
 project=${1}
 secretName="${project}_cloudflare"
 
+cloudflare=$(grep CLOUDFLARE config | cut -f2 -d=)
+echo "MUA $cloudflare"
+case $cloudflare in
+  false | "N" | "NO" | "No" | "no" | "n" )
+    cloudflare=false
+    ;;
+  * )
+    cloudflare=true
+    ;;
+esac
 
-read -p "Please paste your cloudflare api token: " token
-docker secret rm $secretName
-printf $token | docker secret create $secretName -
+echo "ME $cloudflare"
+if [[ $cloudflare == false ]];then
+  sed -i 's/^CLOUDFLARE=.$/CLOUDFLARE=false/g' config
+else
+  sed -i 's/^CLOUDFLARE=.$/CLOUDFLARE=true/g' config
+  read -p "Please paste your cloudflare dns api token: " token
+  docker secret rm $secretName
+  printf $token | docker secret create $secretName -
 
-cat - > /tmp/${project}.secrets.yml<<EOF
+  cat - > /tmp/${project}.secrets.yml<<EOF
 version: '3.7'
 services: 
 
@@ -21,3 +35,6 @@ services:
     secrets:
       - ${secretName}
 EOF
+fi
+
+
